Validacion de usuario al logarser creacion de roles, creacion  de login normal

+++ src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 
+import { getUserSessionServer } from '@/auth/components/actions/auth-action';
 import prisma from '@/lib/prisma'
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
@@ -33,11 +34,15 @@ const postSchema = yup.object({
 });
 
 export async function POST(request: Request) { 
+  const user =  await getUserSessionServer();
+  if (!user) {
+   return NextResponse.json('No autorizado', {status: 401}) 
+  }
 
   try {
     const { complete, description } = await postSchema.validate( await request.json() );
 
-    const todo = await prisma.todo.create({ data: { complete, description } })
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } })
   
     
     return NextResponse.json(todo);
@@ -48,10 +53,14 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) { 
+  const user =  await getUserSessionServer();
+  if (!user) {
+   return NextResponse.json('No autorizado', {status: 401}) 
+  }
 
   try {
 
-    const todo = await prisma.todo.deleteMany({ where: {complete:true}})
+    const todo = await prisma.todo.deleteMany({ where: {complete:true, userId:user.id}})
   
     
     return NextResponse.json('Borrados');
